reset migration

+++ database/migrations/2021_12_07_144518_create_gurus_table.php
@@ -16,7 +16,7 @@ class CreateGurusTable extends Migration
         Schema::create('guru', function (Blueprint $table) {
             $table->id();
             $table->integer('nign')->unique();
-            $table->integer('nip')->nullable()->unique();
+            $table->bigInteger('nip')->nullable()->unique();
             $table->string('nama', 60);
             $table->unsignedBigInteger('jurusan_id');
             $table->timestamps();
